Updated resampling code.
* Changed nearest neighbor interpolator to a faster interger based
one.

* Added a bilinear interpolator (also fast integer based).

The other main benefit of the integer routines are they are
easier to reason about. I'm terribly paranoid about the image
being unrightly shifted a half a pixel or so, but I'm reasonably
sure these routines are correct.

Updated tests for resample module to match output of fast integer

There seems to be a small difference between the integer and floating
point versions, although I did try to match the behaviour (though,
not everything.)