Organize #crop! spec examples
I am about to modify the behavior of crop! so that it takes an optimized
path when it doesn't need to crop the width of the image. Before I do
this, I want to organize these specs a little better to make the diff of
the addition of my new specs easier to see.